new problems and new contests

+++ Codeforces Round 953 (Div. 2)/C.cpp
@@ -1,19 +1,19 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-void cin_array(int a[], int n){
-    for(int i=0; i<n; i++){
+void cin_array(long long int a[], long long int n){
+    for(long long int i=0; i<n; i++){
         cin>>a[i];
     }
 }
 
 int main(){
-    int t;
+    long long int t;
     cin >> t;
     while(t--){
-        int n;
+        long long int n;
         cin >> n;
-        int a[n];
+        long long int a[n];
         cin_array(a, n);
     }
     return 0;
